+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `in_str!` supports byte strings.
+
 ## [0.1.0] - 2024-11-20
 
 ### Added